+++ tools/ros_integration/src/mpsoc_node/include/mpsoc_helper.h
@@ -1,0 +1,8 @@
+#ifndef _MPSOC_HELPER
+#define _MPSOC_HELPER
+
+void add_noc_header(char* buffer);
+
+void dump(char* _mem, uint32_t base, uint32_t length);
+
+#endif /* _MPSOC_HELPER */
