+++ tools/ros_integration/src/mpsoc_node/src/mpsoc_node.cpp
@@ -1,7 +1,8 @@
 #include "ros/ros.h"
-#include "geometry_msgs/Point.h"   // Motion
+#include "sensor_msgs/LaserScan.h"   // Motion
 
 #include "../include/udp_client_server.h"
+#include "../include/mpsoc_helper.h"
 
 #include <iostream>
 
@@ -19,133 +20,72 @@
 //but the waste of memory
 #define UDP_BUFFER_LEN 128
 
+//laser specific definition
+#define HOKUYO_NUM_RANGES 
+
+
 //ros references
 ros::Publisher  pub_mpsoc_out;
 ros::Subscriber pub_mpsoc_in;
 
 //definition of topic type; Include directive
 //must be changed as well
-typedef geometry_msgs::Point topic_t ;
+typedef sensor_msgs::LaserScan topic_t ;
 
 //conection to udp network
 udp_client* uclient;
 udp_server* userver;
 
-//shorthand print
-void dump(char* _mem, uint32_t base, uint32_t length){
-	uint32_t k, l;
-	
-	//mask is necessary to correct a bug(?) when printing
-	//negative hexas.
-	uint32_t mask = 0x000000FF; 
-	int8_t ch;
-	
-	//uint32_t* memptr = (uint32_t*)_mem;
-	//uint32_t  len = _length / 4;
-	for(k = 0; k < length; k += 16){
-		printf("\n%08x ", base + k);
-		for(l = 0; l < 16; l++){
-			printf("%02x ", _mem[k + l] & mask );
-			if (l == 7) putchar(' ');
-		}
-		printf(" |");
-		for(l = 0; l < 16; l++){
-			ch = _mem[k + l];
-			if ((ch >= 32) && (ch <= 126))
-				putchar(ch);
-			else
-				putchar('.');
-		}
-		putchar('|');
-	}
-}
 
-//serialize
-void* pack_data(char* addr, const topic_t::ConstPtr& msg){
-
-	char* ptr = addr;
-	
-	//add x val
-	uint32_t x_val = msg->x * 100000;
-	*(uint32_t*)ptr = x_val;
-	
-	//inc 4 bytes
-	ptr += 4;
-	
-	//add x val
-	uint32_t y_val = msg->y * 100000;
-	*(uint32_t*)ptr = y_val;
-	
-	ptr += 4;
-	
-	//add x val
-	uint32_t z_val = msg->z * 100000;
-	*(uint32_t*)ptr = z_val;
-}
-
-//unserialize
-topic_t unpack_data(char* addr){
-	
-	topic_t t;
-	t.x = *(int32_t*)(&addr[0]);
-	t.y = *(int32_t*)(&addr[4]);
-	t.z = *(int32_t*)(&addr[8]);
-	
-	return t;
-}
+#define MOD_SKIP_PACKET 100
+unsigned int packet_counter = 0;
 
 //Callback that consumes message on the input topic and process
 //information. Message is published back into output topic at
 //the end of processing.
 void mpsoc_in_callback(const topic_t::ConstPtr& msg){
 
-	//buffes must be in the callback, otherwise
-	//data will be overwritter by next iterations
-	char buffer[UDP_BUFFER_LEN];
+	//make sure packet counter is always between 0 and MOD_SKIP_PACKET
+	packet_counter = (packet_counter + 1) % MOD_SKIP_PACKET;
+	
+	//ignore MOD_SKIP_PACKET packets
+	if(packet_counter == 0){
+		
+		//buffes must be in the callback, otherwise
+		//data will be overwritter by next iterations
+		char buffer[UDP_BUFFER_LEN];
 
-	for(int i = 0; i < UDP_BUFFER_LEN; i++)
-		buffer[i] = 0x00;
+		//clean buffer
+		for(int i = 0; i < UDP_BUFFER_LEN; i++) 
+			buffer[i] = 0x00;
 
-	//add noc headers 
-	buffer[0] = 0x11;  //(1,1) is core #5
-	buffer[1] = 0x00; 
-
-	buffer[2] = 0x3e; 
-	buffer[3] = 0x00;  //length flit: 0x3e = 62 flits
-
-	buffer[4] = 0x00;  //payload
-	buffer[5] = 0x05;  //target_cpu (5)
-
-	buffer[6] = 0xe8;	//src_port (5000)
-	buffer[7] = 0x00;  //src_cpu (0,0)
-	
-	buffer[8] = 0x88;  //msg_size
-	buffer[9] = 0x13;  //0x1388 = 5000 dec
-	
-	buffer[10] = 0x64;
-	buffer[11] = 0x00;
-	
-	buffer[12] = 0x01;
-	buffer[16] = 0x20;
-	
-	for(int i = 0; i < 20; i++)
-		buffer[17 + i] = 0x63;
-	
-	buffer[40] = 0x0a; // \n
-	
-	//pack data
-	pack_data(&buffer[20], msg);
-	
-	//send via udp
-	uclient->send((const char*)buffer, UDP_BUFFER_LEN);
-	
-	ROS_INFO("data sent");
-	dump(buffer, 0, UDP_BUFFER_LEN);
-	printf("\n");
-	
-	//do whatever has to be done with the data
-	//start_addr = remove_noc_headers(buffer);
-	
+		//add noc header to the packet
+		add_noc_header(buffer);
+		
+		//add laser info
+		int j = 1;
+		for(float i = msg->angle_min; i < msg->angle_max; i += msg->angle_increment){
+			
+			//ignoring outranged values
+			if(msg->ranges[j] >= msg->range_min && msg->ranges[j] <= msg->range_max){
+				
+				//send 3xx ranges intead of 6xx
+				if(j % 2 == 0){
+				
+					buffer[20] = msg->ranges[j] * 100;
+						
+						
+					uclient->send((const char*)buffer, UDP_BUFFER_LEN);		
+					printf("%d = %f, inc = %f\n", j, i, msg->angle_increment);
+				}
+			}
+			j++;
+			
+		}
+		
+		ROS_INFO("data sent");	
+		
+	}
 }
 
 //Non-blocking function for receiveing packets from 
@@ -165,14 +105,11 @@ void* mpsoc_out(void* v){
 		
 		//unpack data
 		topic_t data;
-		data = unpack_data(start_addr);
 		
 		//publish through ros 
 		pub_mpsoc_out.publish(data);
 		
 		ROS_INFO("data recv'd");
-		dump(buffer, 0, UDP_BUFFER_LEN);
-		printf("\n");
 	}
 }
 
@@ -186,8 +123,8 @@ int main(int argc,char **argv)
 	
 	//ROS_INFO();
 	//subscribe and advertise to input and output topics. 
-	pub_mpsoc_in  = n.subscribe("/mpsoc_in", 1, mpsoc_in_callback); 
-	pub_mpsoc_out = n.advertise<topic_t>("/mpsoc_out", 1);
+	pub_mpsoc_in  = n.subscribe("/hokuyo_laser", 10, mpsoc_in_callback); 
+	pub_mpsoc_out = n.advertise<topic_t>("/mpsoc_out", 10);
 
 	//initialize udp bridge
 	const std::string& client_addr = MPSOC_ADDR;
